Develop

+++ internal/database/bookings.go
@@ -6,11 +6,14 @@ import (
 )
 
 /*
-InsertBooking is a
+InsertBooking will retrieve true after successful insertion.
+In other case, will retrieve false.
 */
 func InsertBooking(id int, people int, state int, table int) bool {
-
-	db := getConnection()
+	db, errCon := getConnection()
+	if errCon != nil {
+		return false
+	}
 	sqlStatement := `INSERT INTO "table-booking-sch"."BOOKINGS"
 	("id", "people", "status", "timestamp_created", "timestamp_last_updated", "table") 
 	VALUES ($1, $2, $3, NOW(), NOW(), $4)`
@@ -24,11 +27,15 @@ func InsertBooking(id int, people int, state int, table int) bool {
 }
 
 /*
-UpdateStatusBookingByID is a
+UpdateStatusBookingByID will retrieve true after successful update.
+In other case, will retrieve false.
 */
 func UpdateStatusBookingByID(id int, newStatus int) bool {
-	db := getConnection()
-	sqlStatement := `UPDATE "table-booking-sch"."BOOKINGS" SET "status" = $1 WHERE id = $2`
+	db, errCon := getConnection()
+	if errCon != nil {
+		return false
+	}
+	sqlStatement := `UPDATE "table-booking-sch"."BOOKINGS" SET "status" = $1 WHERE "id" = $2`
 	_, err := db.Exec(sqlStatement, newStatus, id)
 	defer db.Close()
 	if err != nil {
@@ -39,10 +46,14 @@ func UpdateStatusBookingByID(id int, newStatus int) bool {
 }
 
 /*
-FindBookingByID is a
+FindBookingByID will retrieve the values to build a booking after successful search.
+In other case, will retrieve multiple zeros.
 */
 func FindBookingByID(idToSearch int) (int, int, int, int) {
-	db := getConnection()
+	db, errCon := getConnection()
+	if errCon != nil {
+		return 0, 0, 0, 0
+	}
 	sqlStatement := `SELECT * FROM "table-booking-sch"."BOOKINGS" WHERE "id" = $1`
 	id, people, status, table := 0, 0, 0, 0
 	timeCreated, timeUpdated := time.Now(), time.Now()
@@ -64,20 +75,16 @@ func FindBookingByID(idToSearch int) (int, int, int, int) {
 }
 
 /*
-TruncateBookings is a
+TruncateBookings will retrieve empty error after successful truncate.
+In other case, will retrieve a filled error.
 */
 func TruncateBookings() error {
-	db := getConnection()
+	db, errCon := getConnection()
+	if errCon != nil {
+		return errCon
+	}
 	sqlStatement := `TRUNCATE TABLE "table-booking-sch"."BOOKINGS"`
 	_, err := db.Exec(sqlStatement)
 	defer db.Close()
 	return err
 }
-
-/*
-	INSERT INTO users (age, email, first_name, last_name)
-	VALUES ($1, $2, $3, $4)
-	RETURNING id`
-	id := 0
-	err := db.QueryRow(sqlStatement, 30, "jon@calhoun.io", "Jonathan", "Calhoun").Scan(&id)
-*/
