+++ internal/database/tables.go
@@ -3,11 +3,14 @@ package database
 import "log"
 
 /*
-InsertTable is a
+InsertTable will retrieve true after successful insertion.
+In other case, will retrieve false.
 */
 func InsertTable(id int, seats int, status int) bool {
-
-	db := getConnection()
+	db, errCon := getConnection()
+	if errCon != nil {
+		return false
+	}
 	sqlStatement := `INSERT INTO "table-booking-sch"."TABLES"
 	("id", "seats", "status", "timestamp_created", "timestamp_last_updated") 
 	VALUES ($1, $2, $3, NOW(), NOW())`
@@ -21,10 +24,14 @@ func InsertTable(id int, seats int, status int) bool {
 }
 
 /*
-CheckAvailableTable is a
+CheckAvailableTable will retrieve the table's id after successful search.
+In other case, will retrieve a zero.
 */
 func CheckAvailableTable(requiredSeats int) int {
-	db := getConnection()
+	db, errCon := getConnection()
+	if errCon != nil {
+		return 0
+	}
 	sqlStatement := `SELECT "id" FROM "table-booking-sch"."TABLES" WHERE "status" = 1 AND "seats" >= $1`
 	id := 0
 	rows, err := db.Query(sqlStatement, requiredSeats)
@@ -45,12 +52,16 @@ func CheckAvailableTable(requiredSeats int) int {
 }
 
 /*
-UpdateStatusTableByID is a
+UpdateStatusTableByID will retrieve true after successful update.
+In other case, will retrieve a zero.
 */
 func UpdateStatusTableByID(id int, newStatus int) bool {
-	db := getConnection()
-	sqlStatement := `UPDATE "table-booking-sch"."TABLES" SET "status" = $1 WHERE id = $2`
-	err := db.QueryRow(sqlStatement, newStatus, id)
+	db, errCon := getConnection()
+	if errCon != nil {
+		return false
+	}
+	sqlStatement := `UPDATE "table-booking-sch"."TABLES" SET "status" = $1 WHERE "id" = $2`
+	_, err := db.Exec(sqlStatement, newStatus, id)
 	defer db.Close()
 	if err != nil {
 		log.Println(err)
@@ -60,10 +71,14 @@ func UpdateStatusTableByID(id int, newStatus int) bool {
 }
 
 /*
-TruncateTables is a
+TruncateTables will retrieve empty error after successful truncate.
+In other case, will retrieve a filled error.
 */
 func TruncateTables() error {
-	db := getConnection()
+	db, errCon := getConnection()
+	if errCon != nil {
+		return errCon
+	}
 	sqlStatement := `TRUNCATE TABLE "table-booking-sch"."TABLES"`
 	_, err := db.Exec(sqlStatement)
 	defer db.Close()
